pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Async API Sprint 1"
authors = ["nizdolin <nizdolin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.3"
elasticsearch = "^8.12.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.api.dependencies]
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
redis = "^5.0.1"
uvloop = "^0.19.0"
orjson = "^3.9.12"
aiohttp = "^3.9.1"

[tool.poetry.group.etl.dependencies]
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^6.1.0"
wemake-python-styleguide = "^0.18.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pyright = "^1.1.347"
pyproject-flake8 = "^6.1.0"


[tool.poetry.group.prod.dependencies]
gunicorn = "^21.2.0"

[tool.black]
line-length = 120
exclude = ".venv"

[tool.flake8]
inline-quotes = "\""
format = "wemake"
show-source = true
statistics = "false"
doctests = true
select = "C,E,F,W,B,B950"
max-try-body-length = 10
max-local-variables = 10
# Plugins:
max-complexity = 6
max-line-length = 120
# darglint configuration:
# https://github.com/terrencepreilly/darglint
strictness = "long"
docstring-style = "numpy"
# Self settings:
max-imports = 17
# Excluding some directories:
exclude = ["**/__pycache__", ".venv", ".chroma", "flake-report"]

[tool.isort]
profile = "wemake"
default_section = "THIRDPARTY"
known_first_party = "wemake_python_styleguide*"
skip_glob = ["**/__pycache__", ".venv", ".chroma", "flake-report"]

[tool.pyright]
include = ["*"]
exclude = ["**/__pycache__", ".venv", ".chroma"]
defineConstant = { DEBUG = true }
venvPath = "./"
venv = ".venv"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "Linux"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
